Add Inline Tag support.
As discussed at the PR https://github.com/visionmedia/jade/pull/1323
Jade now can embed tags "inline" in plain text, using the delimiters
"#[" and "]". For example:

  a(href='/') #[i Emphasized] Home Page

becomes:

  <a href="/"><i>Emphasized</i> Home Page</a>

This is quite useful!

I made two new syntax matches, jadeTagInlineText and jadeTagBlockChar,
and three new syntax regions: jadePipedText, jadeTagBlockText, and
jadeTextInlineJade. Each of the three "Text" items "contains" the new
jadeTextInlineJade region as well as the pre-existing jadeInterpolation
region.

I also added a "\+" to the regex for jadeBlockExpansionChar, so that it
can handle more than one space between the colon and the next tag. This
is consistent with Jade's behavior.