Indent on trailing "." for tag-block.
Adding `jadeTagBlockChar`, which is just a "." at the end of a tag line,
to the list of syntax groups that cause an increase in indentation. This
is appropriate because this forms a "tag block". Tag blocks are the
easiest way to include large blocks of "plain" text in the html element
generated by the tag:

```jade
  p.
    Here is a big block of text that will be inside the paragraph
    element and you will note that we didn't even have to use any of
    those funny-looking pipes!
```